提交

+++ os/src/task/manager.rs
@@ -26,7 +26,7 @@ impl TaskManager {
     /// Take a process out of the ready queue 调度算法 我觉得可以改成堆 下班回家试一试
     pub fn fetch(&mut self) -> Option<Arc<TaskControlBlock>> {
         let mut min_idx=0;
-        let mut min_stride=i32::MAX;
+        let mut min_stride=usize::MAX;
         for (idx,task) in self.ready_queue.iter().enumerate(){
             let inner=task.inner_exclusive_access();
             if inner.task_status==TaskStatus::Ready&&inner.stride<min_stride{
